feat(index): 重构后台首页仪表盘、新增待办统计并修复关联查询问题（v5.1.1）

+++ app/common.php
@@ -1277,7 +1277,7 @@ function cpDecode($string,$key=''){
         $box[$j] = $tmp;
         $result .= chr(ord($string[$i]) ^ ($box[($box[$a] + $box[$j]) % 256]));
     }
-    if((substr($result, 0, 10) == 0 || substr($result, 0, 10) - time() > 0) && substr($result, 10, 16) == substr(md5(substr($result, 26).$keyb), 0, 16)) {
+    if(((int)substr($result, 0, 10) == 0 || (int)substr($result, 0, 10) - time() > 0) && substr($result, 10, 16) == substr(md5(substr($result, 26).$keyb), 0, 16)) {
         return unserialize(substr($result, 26));
     }else{
         return '';
